refactor(cli): extract mcp banner/args/responses/schemas helpers (issue #70)
Renamed mcp.rs to mcp/_legacy.rs as a temporary container; helpers drained
into sibling modules (banner, responses, args, schemas). _legacy.rs will be
deleted once empty (PR1 Tasks 5-8).
